moved cards loader to separate component
